# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: مصنع نبع صافيا - نظام إدارة المخزن والمبيعات
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/dashboard:
    get:
      operationId: getDashboard
      tags: [reports]
      summary: لوحة المتابعة - ملخص شامل
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /reports/income-statement:
    get:
      operationId: getIncomeStatement
      tags: [reports]
      summary: قائمة الدخل (الأرباح والخسائر)
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeStatement"
  /reports/balance-sheet:
    get:
      operationId: getBalanceSheet
      tags: [reports]
      summary: الميزانية العمومية
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheet"
  /reports/rep-sales:
    get:
      operationId: getRepSalesReport
      tags: [reports]
      summary: مبيعات المندوبين
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepSalesSummary"
  /reports/rep-collections:
    get:
      operationId: getRepCollectionsReport
      tags: [reports]
      summary: تحصيلات المندوبين
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepCollectionsSummary"
  /reports/vat:
    get:
      operationId: getVatReport
      tags: [reports]
      summary: تقرير ضريبة القيمة المضافة
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VatReport"
  /reports/monthly-sales:
    get:
      operationId: getMonthlySalesChart
      tags: [reports]
      summary: مبيعات شهرية للرسم البياني
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyChartPoint"
components:
  schemas:
...
    Purchase:
      type: object
      required: [id, invoiceNumber, date, supplierName, items, totalAmount, vatAmount, grandTotal, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        supplierName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        totalAmount:
          type: number
        vatAmount:
          type: number
        grandTotal:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseItem:
      type: object
      required: [productId, productName, quantity, unitPrice, vatRate, subtotal]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        vatRate:
          type: number
        subtotal:
          type: number
...
    Sale:
      type: object
      required: [id, orderNumber, date, customerId, customerName, repId, repName, items, totalAmount, vatAmount, grandTotal, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        date:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        repId:
          type: integer
        repName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        totalAmount:
          type: number
        vatAmount:
          type: number
        grandTotal:
          type: number
        paymentMethod:
          type: string
          enum: [coupon, cash, network]
        couponBookId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, vatRate, subtotal]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        vatRate:
          type: number
        subtotal:
          type: number
...
    InventoryItem:
      type: object
      required: [productId, productName, color, quantity]
      properties:
        productId:
          type: integer
        productName:
          type: string
        color:
          type: string
        quantity:
          type: integer
...
    DashboardData:
      type: object
      required: [totalSalesAmount, totalPurchasesAmount, totalExpenses, netProfit, cashBalance, totalCustomers, totalReps, inventoryItems, recentSales, recentPurchases, monthlySales, topReps]
      properties:
        totalSalesAmount:
          type: number
        totalPurchasesAmount:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        cashBalance:
          type: number
        totalCustomers:
          type: integer
        totalReps:
          type: integer
        inventoryItems:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        recentSales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        recentPurchases:
          type: array
          items:
            $ref: "#/components/schemas/Purchase"
        monthlySales:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyChartPoint"
        topReps:
          type: array
          items:
            $ref: "#/components/schemas/RepSalesSummary"
    MonthlyChartPoint:
      type: object
      required: [month, year, label, salesAmount, purchasesAmount, expenses, profit]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        salesAmount:
          type: number
        purchasesAmount:
          type: number
        expenses:
          type: number
        profit:
          type: number
    IncomeStatement:
      type: object
      required: [from, to, totalRevenue, totalCost, grossProfit, totalExpenses, netProfit, vatCollected, vatPaid, vatBalance, revenueBreakdown, expenseBreakdown]
      properties:
        from:
          type: string
        to:
          type: string
        totalRevenue:
          type: number
        totalCost:
          type: number
        grossProfit:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        vatCollected:
          type: number
        vatPaid:
          type: number
        vatBalance:
          type: number
        revenueBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownItem"
        expenseBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownItem"
    BalanceSheet:
      type: object
      required: [date, totalAssets, totalLiabilities, equity, assets, liabilities]
      properties:
        date:
          type: string
        totalAssets:
          type: number
        totalLiabilities:
          type: number
        equity:
          type: number
        assets:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownItem"
        liabilities:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownItem"
    BreakdownItem:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: number
    RepSalesSummary:
      type: object
      required: [repId, repName, area, totalOrders, totalAmount, totalVat, grandTotal]
      properties:
        repId:
          type: integer
        repName:
          type: string
        area:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalAmount:
          type: number
        totalVat:
          type: number
        grandTotal:
          type: number
    RepCollectionsSummary:
      type: object
      required: [repId, repName, area, cashCollected, couponCollected, networkCollected, total]
      properties:
        repId:
          type: integer
        repName:
          type: string
        area:
          type: ["string", "null"]
        cashCollected:
          type: number
        couponCollected:
          type: number
        networkCollected:
          type: number
        total:
          type: number
    VatReport:
      type: object
      required: [from, to, companyName, vatNumber, salesVat, purchasesVat, netVat, salesDetails, purchasesDetails]
      properties:
        from:
          type: string
        to:
          type: string
        companyName:
          type: string
        vatNumber:
          type: string
        salesVat:
          type: number
        purchasesVat:
          type: number
        netVat:
          type: number
        salesDetails:
          type: array
          items:
            $ref: "#/components/schemas/VatDetail"
        purchasesDetails:
          type: array
          items:
            $ref: "#/components/schemas/VatDetail"
    VatDetail:
      type: object
      required: [date, reference, description, baseAmount, vatAmount, total]
      properties:
        date:
          type: string
        reference:
          type: string
        description:
          type: string
        baseAmount:
          type: number
        vatAmount:
          type: number
        total:
          type: number